[refactor] render 와 관련된 함수를 컴포넌트에서 분리 - getData를 사용하는 render 함수들을 render 폴더 내부의 components 파일로 이동했다.
