+++ public/JS/components/container/container.js
@@ -1,15 +1,4 @@
-import {renderSelectDayHome} from './selectDay/selectDayList.js';
-import {renderRankingList} from './rankingList.js';
-import {makeWebtoonList} from './webtoonList.js';
-import {makeEventSection} from './eventSection.js';
-import {getData} from '../utility.js';
-
-function renderContainer(containerInfo, tab) {
-  return makeContainerLayout(containerInfo, tab)
-  .then(containerLayout => makeContainer(containerInfo, containerLayout))
-}
-
-function makeContainer(containerInfo, containerLayout) {
+export default function makeContainer(containerInfo, containerLayout) {
   return `
     <div class="container ${containerInfo.class}">
       ${makeContainerTitle(containerInfo.title)}
@@ -29,27 +18,3 @@ function makeContainerTitle(title) {
   </div>
   `
 }
-
-// ==================== layout ====================
-function makeContainerLayout(containerInfo, tab) {
-  switch (containerInfo.layout) {
-    case 'SMALL_CARD_LAYOUT':
-      return containerInfo.datasetTop === 'daily'
-        ? renderSelectDayHome(containerInfo.items)
-        : renderGenreTop(containerInfo.items);
-    case 'RANKING_LAYOUT':
-      return renderRankingList(tab);
-      // case 'EVENT_LAYOUT': 
-      // return makeEventSection(containerInfo.class);
-      // makeEventSection는 프로미스가 아니라 문자열을 반환하기 때문에 아무일도 안일어난다.
-  }
-}
-
-function renderGenreTop(items) {
-  const url = 'http://localhost:3000/home-genre-top'
-  return getData(url)
-    .then(json => json['romanceTop'])
-    .then(data => makeWebtoonList(items, data));
-}
-
-export {renderContainer}
